requirements.txt
py-wifi-helper>=1.0.0
paho-mqtt>=1.6.1
structlog>=23.1.0
rich>=13.5.2
